Commit v3

+++ src/LearningManagementSystem/mainManagementSystem/activities/Actividad.java
@@ -158,10 +158,8 @@ public abstract class Actividad implements Cloneable {
         this.estado = "NO EXITOSA";
     }
 
-    public void addReview(String contenido, Double rating, String nombreUsuario) {
+    public void addReview(String contenido, int rating, String nombreUsuario) {
         Review review = new Review(contenido, rating, nombreUsuario);
-
-
         
         reviews.add(review);
     }
